fix :: json 노출 알람 전부 막음
